Revert "Pro 4025 upgrade Spring boot libraries"
