+++ build.gradle
@@ -2,18 +2,18 @@ import java.util.stream.Collectors
 
 buildscript {
     dependencies {
-        classpath group: 'org.sonarsource.scanner.gradle', name: 'sonarqube-gradle-plugin', version: '2.6.2'
-        classpath("net.serenity-bdd:serenity-gradle-plugin:2.0.30")
+        classpath group: 'org.sonarsource.scanner.gradle', name: 'sonarqube-gradle-plugin', version: '2.5'
+        classpath("net.serenity-bdd:serenity-gradle-plugin:1.5.2")
     }
 }
 
 plugins {
     id 'application'
     id 'pmd'
-    id 'org.springframework.boot' version '2.0.7.RELEASE'
-    id 'org.owasp.dependencycheck' version '4.0.2'
-    id 'com.github.ben-manes.versions' version '0.20.0'
-    id 'com.gorylenko.gradle-git-properties' version '1.5.2'
+    id 'org.springframework.boot' version '1.5.14.RELEASE'
+    id 'org.owasp.dependencycheck' version '3.2.1'
+    id 'com.github.ben-manes.versions' version '0.15.0'
+    id 'com.gorylenko.gradle-git-properties' version '1.4.21'
     id 'info.solidsoft.pitest' version '1.3.0'
 }
 
@@ -158,25 +158,24 @@ processResources {
 }
 
 def versions = [
-        reformsJavaLoggingVersion: '3.0.4',
-        springBootVersion: '2.0.7.RELEASE',
-        serenityVersion: '2.0.30',
+        logging: '2.2.1'
 ]
 
 dependencies {
-    compile group: 'org.springframework.boot', name: 'spring-boot-starter-actuator', version: versions.springBootVersion
-    compile group: 'org.springframework.boot', name: 'spring-boot-starter-data-rest', version: versions.springBootVersion
-    compile group: 'org.springframework.boot', name: 'spring-boot-starter-data-jpa', version: versions.springBootVersion
 
-    compile ('org.postgresql:postgresql:42.2.5') {
+    compile ('org.springframework.boot:spring-boot-starter-actuator')
+    compile ('org.springframework.boot:spring-boot-starter-data-rest')
+    compile ('org.springframework.boot:spring-boot-starter-data-jpa')
+    
+    compile ('org.postgresql:postgresql:42.2.2') {
         exclude module: 'slf4j-simple'
     }
 
-    compile group: 'org.liquibase', name: 'liquibase-core', version: '3.6.2'
+    compile group: 'org.liquibase', name: 'liquibase-core', version: '3.6.1'
 
-    compile group: 'uk.gov.hmcts.reform', name: 'java-logging', version: versions.reformsJavaLoggingVersion
-    compile group: 'uk.gov.hmcts.reform', name: 'java-logging-appinsights', version: versions.reformsJavaLoggingVersion
-    compile group: 'uk.gov.hmcts.reform', name: 'java-logging-httpcomponents', version: versions.reformsJavaLoggingVersion
+    compile group: 'uk.gov.hmcts.reform', name:'java-logging', version: versions.logging
+    compile group: 'uk.gov.hmcts.reform', name: 'java-logging-appinsights', version: versions.logging
+    compile group: 'uk.gov.hmcts.reform', name: 'java-logging-httpcomponents', version: versions.logging
     compile group: 'org.springframework.hateoas', name: 'spring-hateoas'
 
     testCompile group: 'org.pitest', name: 'pitest', version: '1.3.2'
@@ -185,20 +184,17 @@ dependencies {
 
     testCompile group: 'com.google.code.findbugs', name: 'findbugs', version: '3.0.1'
 
-    testCompile group: 'org.springframework.boot', name: 'spring-boot-starter-test', version: versions.springBootVersion
-
+    testCompile group: 'org.springframework.boot', name: 'spring-boot-starter-test'
     testCompile group: 'com.h2database', name: 'h2', version: '1.4.197'
-    testCompile group: 'io.rest-assured', name: 'rest-assured', version: '3.0.7'
 
-    testFunctionalCompile group: 'org.springframework.boot', name: 'spring-boot-starter-test', version: versions.springBootVersion
-    testFunctionalCompile group: 'net.serenity-bdd', name: 'serenity-junit', version: versions.serenityVersion
-    testFunctionalCompile group: 'net.serenity-bdd', name: 'serenity-core', version: versions.serenityVersion
-    testFunctionalCompile group: 'net.serenity-bdd', name: 'serenity-rest-assured', version: versions.serenityVersion
-    testFunctionalCompile group: 'net.serenity-bdd', name: 'serenity-spring', version: versions.serenityVersion
-    testFunctionalCompile group: 'io.jsonwebtoken', name: 'jjwt', version: '0.9.1'
-    testFunctionalCompile group: 'org.apache.commons', name: 'commons-lang3', version: '3.8.1'
+    testFunctionalCompile group: 'net.serenity-bdd', name: 'serenity-junit', version: '1.5.2'
+    testFunctionalCompile group: 'net.serenity-bdd', name: 'serenity-core', version: '1.5.2'
+    testFunctionalCompile group: 'net.serenity-bdd', name: 'serenity-rest-assured', version: '1.5.2'
+    testFunctionalCompile group: 'net.serenity-bdd', name: 'serenity-spring', version: '1.0.26'
+    testFunctionalCompile group: 'io.jsonwebtoken', name: 'jjwt', version: '0.9.0'
     testFunctionalCompile sourceSets.main.runtimeClasspath
 
+    testCompile group: 'io.rest-assured', name: 'rest-assured', version: '3.0.7'
     testSmokeCompile sourceSets.main.runtimeClasspath
     testSmokeCompile sourceSets.test.runtimeClasspath
 }
@@ -207,7 +203,7 @@ compileJava {
     options.compilerArgs << '-parameters'
 }
 
-bootJar {
+jar {
     archiveName = 'persistence-service.jar'
     manifest {
