+++ config/owasp/suppressions.xml
@@ -1,44 +1,200 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <suppressions xmlns="https://jeremylong.github.io/DependencyCheck/dependency-suppression.1.1.xsd">
-    <suppress>
-        <notes><![CDATA[
-   file name: javax.annotation-api-1.3.2.jar
-   ]]></notes>
-        <gav regex="true">^javax\.annotation:javax\.annotation-api:.*$</gav>
-        <cpe>cpe:/a:oracle:glassfish</cpe>
-    </suppress>
 
-    <suppress>
-        <notes><![CDATA[
-   file name: javax.annotation-api-1.3.2.jar
+  <suppress>
+    <notes><![CDATA[
+   file name: spring-boot-starter-data-rest-1.5.13.RELEASE.jar contains spring_data_rest:2.6.12.RELEASE
    ]]></notes>
-        <gav regex="true">^javax\.annotation:javax\.annotation-api:.*$</gav>
-        <cve>CVE-2013-2566</cve>
-    </suppress>
+    <gav regex="true">^org\.springframework\.boot:spring-boot-starter-data-rest:.*$</gav>
+    <cpe>cpe:/a:pivotal_software:spring_data_rest</cpe>
+  </suppress>
 
-    <suppress>
-        <notes><![CDATA[
-   file name: javax.annotation-api-1.3.2.jar
+  <suppress>
+    <notes><![CDATA[
+   file name: spring-boot-starter-data-rest-1.5.13.RELEASE.jar contains spring-data-commons:1.13.12.RELEASE
    ]]></notes>
-        <gav regex="true">^javax\.annotation:javax\.annotation-api:.*$</gav>
-        <cve>CVE-2015-2808</cve>
-    </suppress>
+    <gav regex="true">^org\.springframework\.boot:spring-boot-starter-data-rest:.*$</gav>
+    <cpe>cpe:/a:pivotal_software:spring_boot</cpe>
+  </suppress>
 
-    <suppress>
-        <notes><![CDATA[
-    file name: dom4j-1.6.1.jar
+  <suppress>
+    <notes><![CDATA[
+   file name: tomcat-annotations-api-8.5.31.jar
+   https://github.com/spring-projects/spring-boot/issues/11110
+   ]]></notes>
+    <gav regex="true">^org\.apache\.tomcat:tomcat-annotations-api:.*$</gav>
+    <cve>CVE-2016-5425</cve>
+  </suppress>
+
+  <suppress>
+    <notes><![CDATA[
+   file name: tomcat-annotations-api-8.5.31.jar
+   https://github.com/spring-projects/spring-boot/issues/11110
+   ]]></notes>
+    <gav regex="true">^org\.apache\.tomcat:tomcat-annotations-api:.*$</gav>
+    <cve>CVE-2016-6325</cve>
+  </suppress>
+
+  <suppress>
+    <notes><![CDATA[
+   file name: tomcat-annotations-api-8.5.31.jar
+   https://github.com/spring-projects/spring-boot/issues/11110
+   ]]></notes>
+    <gav regex="true">^org\.apache\.tomcat:tomcat-annotations-api:.*$</gav>
+    <cve>CVE-2017-6056</cve>
+  </suppress>
+
+  <suppress>
+    <notes><![CDATA[
+   file name: postgresql-42.2.2.jar
+   N/A - exists in SUSE linux
+   ]]></notes>
+    <gav regex="true">^org\.postgresql:postgresql:.*$</gav>
+    <cve>CVE-2017-14798</cve>
+  </suppress>
+
+  <suppress>
+    <notes><![CDATA[
+   file name: postgresql-42.2.2.jar
+   ]]></notes>
+    <gav regex="true">^org\.postgresql:postgresql:.*$</gav>
+    <cve>CVE-2018-1115</cve>
+  </suppress>
+
+   <suppress>
+    <notes><![CDATA[
+    file name: spring-webmvc-4.3.18.RELEASE.jar
     ]]></notes>
-        <gav regex="true">^dom4j:dom4j:.*$</gav>
-        <cve>CVE-2018-1000632</cve>
-    </suppress>
+    <gav regex="true">^org\.springframework:spring-webmvc:.*$</gav>
+    <cve>CVE-2018-1258</cve>
+  </suppress>
 
-    <suppress>
-        <notes><![CDATA[
+  <suppress>
+    <notes><![CDATA[
+    file name: spring-web-4.3.18.RELEASE.jar
+    ]]></notes>
+    <gav regex="true">^org\.springframework:spring-web:.*$</gav>
+    <cve>CVE-2018-1258</cve>
+  </suppress>
+
+  <suppress>
+    <notes><![CDATA[
+    file name: spring-context-support-4.3.18.RELEASE.jar
+    ]]></notes>
+    <gav regex="true">^org\.springframework:spring-context-support:.*$</gav>
+    <cve>CVE-2018-1258</cve>
+  </suppress>
+
+  <suppress>
+    <notes><![CDATA[
+    file name: spring-context-4.3.18.RELEASE.jar
+    ]]></notes>
+    <gav regex="true">^org\.springframework:spring-context:.*$</gav>
+    <cve>CVE-2018-1258</cve>
+  </suppress>
+
+  <suppress>
+    <notes><![CDATA[
+    file name: spring-aop-4.3.18.RELEASE.jar
+    ]]></notes>
+    <gav regex="true">^org\.springframework:spring-aop:.*$</gav>
+    <cve>CVE-2018-1258</cve>
+  </suppress>
+
+  <suppress>
+    <notes><![CDATA[
+    file name: spring-beans-4.3.18.RELEASE.jar
+    ]]></notes>
+    <gav regex="true">^org\.springframework:spring-beans:.*$</gav>
+    <cve>CVE-2018-1258</cve>
+  </suppress>
+
+  <suppress>
+    <notes><![CDATA[
+    file name: spring-expression-4.3.18.RELEASE.jar
+    ]]></notes>
+    <gav regex="true">^org\.springframework:spring-expression:.*$</gav>
+    <cve>CVE-2018-1258</cve>
+  </suppress>
+
+  <suppress>
+    <notes><![CDATA[
+    file name: spring-core-4.3.18.RELEASE.jar
+    ]]></notes>
+    <gav regex="true">^org\.springframework:spring-core:.*$</gav>
+    <cve>CVE-2018-1258</cve>
+  </suppress>
+
+  <suppress>
+     <notes><![CDATA[
+     file name: spring-aspects-4.3.18.RELEASE.jar
+     ]]></notes>
+     <gav regex="true">^org\.springframework:spring-aspects:.*$</gav>
+     <cpe>cpe:/a:pivotal:spring_framework</cpe>
+  </suppress>
+
+  <suppress>
+     <notes><![CDATA[
+     file name: spring-orm-4.3.18.RELEASE.jar
+     ]]></notes>
+     <gav regex="true">^org\.springframework:spring-orm:.*$</gav>
+     <cve>CVE-2018-1258</cve>
+  </suppress>
+  <suppress>
+   <notes><![CDATA[
+   file name: postgresql-42.2.2.jar
+   ]]></notes>
+   <gav regex="true">^org\.postgresql:postgresql:.*$</gav>
+   <cve>CVE-2016-7048</cve>
+  </suppress>
+
+  <suppress>
+     <notes><![CDATA[
+     file name: spring-jdbc-4.3.18.RELEASE.jar
+     ]]></notes>
+     <gav regex="true">^org\.springframework:spring-jdbc:.*$</gav>
+     <cve>CVE-2018-1258</cve>
+  </suppress>
+
+  <suppress>
+     <notes><![CDATA[
+     file name: spring-tx-4.3.18.RELEASE.jar
+     ]]></notes>
+     <gav regex="true">^org\.springframework:spring-tx:.*$</gav>
+     <cve>CVE-2018-1258</cve>
+  </suppress>
+
+  <suppress>
+     <notes><![CDATA[
+     file name: spring-aspects-4.3.18.RELEASE.jar
+     ]]></notes>
+     <gav regex="true">^org\.springframework:spring-aspects:.*$</gav>
+     <cve>CVE-2018-1258</cve>
+  </suppress>
+
+  <suppress>
+   <notes><![CDATA[
+   file name: postgresql-42.2.2.jar
+   ]]></notes>
+   <gav regex="true">^org\.postgresql:postgresql:.*$</gav>
+   <cve>CVE-2016-7048</cve>
+  </suppress>
+
+  <suppress>
+   <notes><![CDATA[
+   file name: dom4j-1.6.1.jar
+   ]]></notes>
+   <gav regex="true">^dom4j:dom4j:.*$</gav>
+   <cve>CVE-2018-1000632</cve>
+  </suppress>
+
+  <suppress>
+   <notes><![CDATA[
    file name: applicationinsights-core-2.0.2.jar/META-INF/maven/com.google.guava/guava/pom.xml
    ]]></notes>
-        <gav regex="true">^com\.google\.guava:guava:.*$</gav>
-        <cpe>cpe:/a:google:guava</cpe>
-    </suppress>
+   <gav regex="true">^com\.google\.guava:guava:.*$</gav>
+   <cpe>cpe:/a:google:guava</cpe>
+  </suppress>
 
     <suppress>
         <notes><![CDATA[
