+++ src/main/java/uk/gov/hmcts/probate/services/persistence/repository/InvitedataRepository.java
@@ -8,12 +8,11 @@ import org.springframework.data.rest.core.annotation.RestResource;
 import uk.gov.hmcts.probate.services.persistence.model.Invitedata;
 
 import java.util.List;
-import java.util.Optional;
 
 @RepositoryRestResource(collectionResourceRel = "invitedata", path = "invitedata")
 public interface InvitedataRepository extends PagingAndSortingRepository<Invitedata, String> {
 
-    Optional<Invitedata> findById(@Param("id") String id);
+    Invitedata findById(@Param("id") String id);
 
     @RestResource(path = "formdata")
     List<Invitedata> findByFormdataId(@Param("id") String id);
