+++ src/main/java/uk/gov/hmcts/probate/services/persistence/services/SequenceNumberService.java
@@ -1,7 +1,8 @@
 package uk.gov.hmcts.probate.services.persistence.services;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.jdbc.support.incrementer.PostgresSequenceMaxValueIncrementer;
+import org.springframework.boot.context.config.ResourceNotFoundException;
+import org.springframework.jdbc.support.incrementer.PostgreSQLSequenceMaxValueIncrementer;
 import org.springframework.stereotype.Service;
 import uk.gov.hmcts.probate.services.persistence.model.RegistryNotConfiguredException;
 
@@ -11,7 +12,7 @@ import java.util.Map;
 public class SequenceNumberService {
 
     @Autowired
-    private Map<String, PostgresSequenceMaxValueIncrementer> registrySequenceNumbers;
+    private Map<String, PostgreSQLSequenceMaxValueIncrementer> registrySequenceNumbers;
 
     public Long getNext(String registryName) {
