+++ src/main/resources/application.properties
@@ -6,11 +6,5 @@ probate.postgresql.password=password123
 probate.persistence.port=8282
 probate.persistence.show_sql=true
 management.info.git.mode=full
-management.endpoints.web.base-path=/
 spring.info.git.location=classpath:uk/gov/hmcts/probate/services/persistence/git.properties
 
-# https://vkuzel.com/spring-boot-jpa-hibernate-atomikos-postgresql-exception
-# Disable feature detection by this undocumented parameter. Check the org.hibernate.engine.jdbc.internal.JdbcServiceImpl.configure method for more details.
-spring.jpa.properties.hibernate.temp.use_jdbc_metadata_defaults = false
-# Because detection is disabled you have to set correct dialect by hand.
-spring.jpa.database-platform=org.hibernate.dialect.PostgreSQL9Dialect
