+++ src/test/java/uk/gov/hmcts/probate/services/persistence/repository/InvitedataRepositoryTest.java
@@ -9,7 +9,7 @@ import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.jdbc.EmbeddedDatabaseConnection;
+import org.springframework.boot.autoconfigure.jdbc.EmbeddedDatabaseConnection;
 import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
 import org.springframework.boot.test.context.SpringBootTest;
@@ -36,6 +36,9 @@ public class InvitedataRepositoryTest {
   @Autowired
   private MockMvc mockMvc;
 
+  @Autowired
+  private InvitedataRepository invitedataRepository;
+
   @Before
   public void setUp() throws Exception {
     mockMvc.perform(post("/invitedata/")
@@ -47,6 +50,12 @@ public class InvitedataRepositoryTest {
   @Test
   public void shouldGetInviteData() throws Exception {
     mockMvc.perform(get("/invitedata/1"))
+        .andExpect(status().isOk());
+  }
+
+  @Test
+  public void shouldFindById() throws Exception {
+    mockMvc.perform(get("/invitedata/search/findById?id=1"))
         .andExpect(status().isOk());
   }
