+++ src/test/resources/application.properties
@@ -7,5 +7,3 @@ probate.postgresql.database=probate
 probate.postgresql.user=probate
 probate.postgresql.password=password123
 probate.persistence.show_sql=true
-
-management.endpoints.web.base-path=/
